wip: move all cli related code to new package. clean up deps
