+++ packages/plugin-docs-cli/package.json
@@ -6,9 +6,7 @@
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "bin": {
-    "plugin-docs": "./dist/bin/run.js"
-  },
+  "bin": "./dist/bin/run.js",
   "files": [
     "dist",
     "LICENSE",
@@ -49,5 +47,28 @@
   },
   "engines": {
     "node": ">=24"
+  },
+  "dependencies": {
+    "@grafana/plugin-docs-renderer": "^0.0.1",
+    "chokidar": "^3.6.0",
+    "debug": "^4.3.7",
+    "ejs": "^4.0.1",
+    "express": "^4.18.0",
+    "fs-extra": "^11.3.3",
+    "github-slugger": "^1.5.0",
+    "globby": "^11.1.0",
+    "gray-matter": "^4.0.3",
+    "marked": "^13.0.0",
+    "minimist": "^1.2.8"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.12",
+    "@types/ejs": "^3.1.5",
+    "@types/express": "^4.17.21",
+    "@types/fs-extra": "^11.0.4",
+    "@types/github-slugger": "^1.3.0",
+    "@types/minimist": "^1.2.5",
+    "@types/supertest": "^6.0.2",
+    "supertest": "^7.0.0"
   }
 }
