+++ packages/plugin-docs-cli/src/__fixtures__/test-docs/advanced.md
@@ -1,0 +1,9 @@
+---
+title: Advanced Topics
+description: Advanced topics for experienced users
+sidebar_position: 3
+---
+
+# Advanced Topics
+
+This covers advanced topics.
