+++ packages/plugin-docs-cli/src/__fixtures__/test-docs/config/database.md
@@ -1,0 +1,9 @@
+---
+title: Database
+description: Database configuration options
+sidebar_position: 2
+---
+
+# Database
+
+Configure database connections.
