+++ packages/plugin-docs-cli/src/__fixtures__/test-docs/config/settings.md
@@ -1,0 +1,9 @@
+---
+title: Settings
+description: Configuration settings
+sidebar_position: 1
+---
+
+# Settings
+
+Configure your plugin settings here.
