+++ packages/plugin-docs-cli/src/__fixtures__/test-docs/guide.md
@@ -1,0 +1,9 @@
+---
+title: User Guide
+description: A guide page for users
+sidebar_position: 2
+---
+
+# User Guide
+
+This is a guide page.
