+++ packages/plugin-docs-cli/src/__fixtures__/test-docs/home.md
@@ -1,0 +1,9 @@
+---
+title: Home Page
+description: Welcome to the test docs
+sidebar_position: 1
+---
+
+# Welcome
+
+This is the home page of the test documentation.
