+++ packages/plugin-docs-cli/src/lib/scanner.test.ts
