+++ packages/plugin-docs-cli/src/server/styles/docs.css
