+++ packages/plugin-docs-cli/src/server/views/docs-layout.ejs
