+++ packages/plugin-docs-cli/src/server/views/partials/navigation.ejs
