+++ packages/plugin-docs-cli/src/server/views/partials/toc.ejs
