+++ packages/plugin-docs-renderer/package.json
@@ -2,11 +2,10 @@
   "name": "@grafana/plugin-docs-renderer",
   "version": "0.0.1",
   "private": "true",
-  "description": "A library for rendering Grafana plugin documentation from markdown files.",
+  "description": "A lightweight library for parsing and rendering Grafana plugin documentation from markdown to HTML.",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "bin": "./dist/bin/run.js",
   "files": [
     "dist",
     "LICENSE",
@@ -26,7 +25,6 @@
   },
   "scripts": {
     "build": "rollup -c ../../rollup.config.ts --configPlugin esbuild",
-    "postbuild": "cp -r src/server/views dist/server/ && cp -r src/server/styles dist/server/",
     "dev": "rollup -c ../../rollup.config.ts --configPlugin esbuild --watch",
     "lint": "eslint --cache ./src",
     "lint:fix": "npm run lint -- --fix",
@@ -48,28 +46,14 @@
     "node": ">=24"
   },
   "dependencies": {
-    "@types/ejs": "^3.1.5",
-    "chokidar": "^3.6.0",
-    "debug": "^4.3.7",
-    "ejs": "^4.0.1",
-    "express": "^4.18.0",
-    "fs-extra": "^11.3.3",
     "github-slugger": "^1.5.0",
-    "globby": "^11.1.0",
     "gray-matter": "^4.0.3",
     "isomorphic-dompurify": "^2.16.0",
     "marked": "^13.0.0",
-    "marked-gfm-heading-id": "^4.1.3",
-    "minimist": "^1.2.8"
+    "marked-gfm-heading-id": "^4.1.3"
   },
   "devDependencies": {
-    "@types/debug": "^4.1.12",
     "@types/dompurify": "^3.0.5",
-    "@types/express": "^4.17.21",
-    "@types/fs-extra": "^11.0.4",
-    "@types/github-slugger": "^1.3.0",
-    "@types/minimist": "^1.2.5",
-    "@types/supertest": "^6.0.2",
-    "supertest": "^7.0.0"
+    "@types/github-slugger": "^1.3.0"
   }
 }
