+++ packages/plugin-docs-renderer/src/index.ts
@@ -2,4 +2,4 @@
 // (CLI utilities like filesystem scanning are not exported)
 export { parseMarkdown } from './parser.js';
 export type { ParsedMarkdown } from './parser.js';
-export type { Manifest, Page, MarkdownFiles, Frontmatter } from './types.js';
+export type { Manifest, Page, MarkdownFiles, Frontmatter, Heading } from './types.js';
